requirements.txt
googletrans==4.0.0rc1
ipython==8.17.2
iso639-lang==2.1.0
pandas==2.1.3
pip-chill==1.0.3
shinyswatch==0.4.1
tensorflow==2.13.0
